Move data fetching out of Console page
